Change README to contain more info
Since this project is comming to an end, I thought that we should have proper README.
I was heavily inspired by Cloudkeeper-ONE README, but I changed a few things, one of
which is project name casing. It is written as Cloudkeeper-AWS not `cloudkeeper-aws`
since it is a name not a command. I find it prettier this way.